pyproject.toml
[project]
name = "sap-lap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4==4.13.5",
    "datasets>=4.1.0",
    "docling==2.51.0",
    "faiss-cpu==1.12.0",
    "langchain==0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-experimental==0.3.4",
    "langchain-google-genai==2.1.10",
    "langchain-groq==0.3.8",
    "langgraph==0.6.7",
    "playwright==1.55.0",
    "pymupdf==1.26.4",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
]
